+++ src/v5/server.rs
@@ -32,7 +32,7 @@ pub struct MqttServer<Io, St, C: ServiceFactory, Cn: ServiceFactory, P: ServiceF
     connect: C,
     srv_control: Cn,
     srv_publish: P,
-    max_size: u32,
+    max_size: usize,
     inflight: usize,
     handshake_timeout: usize,
     disconnect_timeout: usize,
@@ -113,7 +113,7 @@ where
     ///
     /// If max size is set to `0`, size is unlimited.
     /// By default max size is set to `0`
-    pub fn max_size(mut self, size: u32) -> Self {
+    pub fn max_size(mut self, size: usize) -> Self {
         self.max_size = size;
         self
     }
@@ -224,6 +224,7 @@ where
                 self.max_topic_alias,
                 self.handshake_timeout,
             ))
+            .max_size(self.max_size)
             .disconnect_timeout(self.disconnect_timeout)
             .build(factory(publish, control, self.max_topic_alias)),
         )
@@ -252,11 +253,12 @@ where
         ntex::unit_config(
             FactoryBuilder2::new(handshake_service_factory2(
                 connect,
-                self.max_size,
+                self.max_size as u32,
                 self.inflight,
                 self.max_topic_alias,
                 self.handshake_timeout,
             ))
+            .max_size(self.max_size)
             .disconnect_timeout(self.disconnect_timeout)
             .build(factory(publish, control, self.max_topic_alias)),
         )
@@ -265,7 +267,7 @@ where
 
 fn handshake_service_factory<Io, St, C>(
     factory: C,
-    max_size: u32,
+    max_size: usize,
     inflight: usize,
     max_topic_alias: u16,
     handshake_timeout: usize,
@@ -289,7 +291,7 @@ where
                     handshake(
                         conn.codec(mqtt::Codec::new()),
                         service.clone(),
-                        max_size,
+                        max_size as u32,
                         max_topic_alias,
                         inflight,
                     )